Refactor Permission and add tests with repository support
Updated `Permission` entity to use `Long` as `subject` type, added cascade deletion rules, and enabled auditing. Introduced `PermissionRepositoryPort` with comprehensive tests to verify CRUD, referential integrity, concurrency, and audit functionality.

+++ src/main/java/com/tritech/tricore/core/domain/primarykeys/PermissionId.java
@@ -9,32 +9,31 @@ import java.io.Serializable;
 /**
  * Represents a composite key for the `Permission` entity.
  * <p>
- * This class is used to identify a `Permission` uniquely through a combination of the
- * following fields:
+ * This class is used to identify a `Permission` uniquely through a combination
+ * of the following fields:
  * <ul>
- * <li>`subject`: The unique identifier of the user associated with the permission.</li>
- * <li>`groupName`: The name of the group associated with the permission .</li>
- * <li>`levelName`: The name of the level associated with the permission .</li>
+ *     <li>`subject`: The unique identifier of the user associated with the
+ *     permission.</li>
+ *     <li>`groupName`: The name of the group associated with the permission
+ *     .</li>
+ *     <li>`levelName`: The name of the level associated with the permission
+ *     .</li>
  * </ul>
  */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class PermissionId implements Serializable {
-
-	/**
-	 * The unique identifier of the user associated with the permission.
-	 */
-	private Integer subject;
-
-	/**
-	 * The name of the group associated with the permission.
-	 */
-	private String groupName;
-
-	/**
-	 * The name of the level associated with the permission.
-	 */
-	private String levelName;
-
+    /**
+     * The unique identifier of the user associated with the permission.
+     */
+    private Long subject;
+    /**
+     * The name of the group associated with the permission.
+     */
+    private String groupName;
+    /**
+     * The name of the level associated with the permission.
+     */
+    private String levelName;
 }
